backend/hardware/requirements.txt
appdirs==1.4.3
attrs==19.3.0
black==19.10b0
Click==7.0
pathspec==0.7.0
pyserial==3.4
regex==2020.1.8
toml==0.10.0
typed-ast==1.4.1
